- Adaugare functionalitate login, register - Conectarea functionalitatilor cu noua interfata user

+++ src/main/java/ro/ubbcluj/map/socialnetwork/service/UtilizatorService.java
@@ -5,11 +5,13 @@ import ro.ubbcluj.map.socialnetwork.domain.validators.ValidationException;
 import ro.ubbcluj.map.socialnetwork.observer.Observable;
 import ro.ubbcluj.map.socialnetwork.observer.Observer;
 import ro.ubbcluj.map.socialnetwork.repository.Repository;
+import ro.ubbcluj.map.socialnetwork.repository.UserDBRepository;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 public class UtilizatorService implements Observable {
 
@@ -22,27 +24,52 @@ public class UtilizatorService implements Observable {
 
     public void adaugaUtilizator(Utilizator utilizator) throws SQLException {
         repo.findAll().forEach(it -> {
-            if (Objects.equals(it.getId(), utilizator.getId()))
-                throw new ValidationException("Exista un utilizator cu acest ID");
+            if (Objects.equals(it.getUsername(), utilizator.getUsername()))
+                throw new ValidationException("Exista un utilizator cu acest username");
+
         });
+        int k = ((UserDBRepository) repo).countUsers();
+        utilizator.setId((long) (k+1));
         repo.save(utilizator);
         notifyAllObservers();
 
     }
 
-    public void stergeUtilizator(Long id) throws SQLException {
+    public void stergeUtilizator(String username) throws SQLException {
+        Long id = searchID(username);
+        if(id == null){
+            throw new ValidationException("Nu exista utilizatorul");
+        }
         repo.delete(id);
         notifyAllObservers();
     }
 
-    public void updateUtilizator(Utilizator utilizator) throws SQLException {
+    private Long searchID(String username) throws SQLException {
+        for (Utilizator it : repo.findAll()) {
+            if (Objects.equals(it.getUsername(), username)) {
+                return it.getId();
+            }
+        }
 
-        if(repo.findOne(utilizator.getId()).isEmpty())
-            throw new ValidationException("Utilizatorul nu exista");
+        return null;
+    }
 
-        repo.update(utilizator);
+    public void updateUtilizator(String username, String firstName, String lastName) throws SQLException {
+        Utilizator user = findUser(username);
+        user.setFirstName(firstName);
+        user.setLastName(lastName);
+        repo.update(user);
         notifyAllObservers();
-  }
+    }
+
+    public Utilizator findUser(String username) throws SQLException {
+        for (Utilizator it : repo.findAll()) {
+            if (Objects.equals(it.getUsername(), username)) {
+                return it;
+            }
+        }
+        return null;
+    }
 
     public List<Utilizator> getAll() throws SQLException {
 
@@ -68,4 +95,15 @@ public class UtilizatorService implements Observable {
             observer.update();
         }
     }
+
+    public boolean verifyLogin(String username, String password) throws SQLException {
+        Long idUser = searchID(username);
+        Optional<Utilizator> user = repo.findOne(idUser);
+        if (user.isPresent()) {
+            return Objects.equals(user.get().getPassword(), password);
+        } else {
+            throw new ValidationException("Utilizatorul nu exista");
+        }
+    }
+
 }
